# public/sdk/openapi.yaml
openapi: "3.0.3"
info:
  title: Print Hub API
  description: |
    Centralized print management middleware API.

    Print Hub decouples client applications from physical printers. Client apps
    submit print jobs via this REST API, and Print Hub routes them to the
    correct print agent (TrayPrint) for execution.

    ## Authentication

    Two authentication modes are supported:

    | Role          | Method    | Header                             |
    |---------------|-----------|------------------------------------|
    | Client App    | API Key   | `X-API-Key: {your_api_key}`        |
    | Print Agent   | Bearer    | `Authorization: Bearer {agent_key}`|

    Rate limits:
    - Client API: 60 requests/minute
    - Agent API:  120 requests/minute
  version: "1.0.0"
  contact:
    name: Print Hub Support
    url: https://print-hub.example.com/support
  license:
    name: Proprietary
servers:
  - url: https://print-hub.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /branches:
    get:
      summary: List all branches
      description: Returns a list of all active company branches.
      operationId: listBranches
      tags: [Discovery]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchListResponse"
...
  /templates:
    get:
      summary: List all print templates
      description: Returns all available template definitions.
      operationId: listTemplates
      tags: [Templates]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
  /templates/{name}:
    get:
      summary: Get template details
      description: Returns detailed information for a specific template including field positions.
      operationId: getTemplate
      tags: [Templates]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Template name (URL-encoded)
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /documents/upload:
    post:
      summary: Upload a document
      operationId: uploadDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, TXT, images)
                label:
                  type: string
                  description: Optional label for the document
              required:
                - file
      responses:
        "200":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /documents:
    get:
      summary: List documents
      operationId: listDocuments
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
  /documents/{id}:
    get:
      summary: Get document metadata
      operationId: getDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Document ID
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a document
      operationId: deleteDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Document deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client app API key obtained from Print Hub admin panel
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "NOT_FOUND"
            message:
              type: string
              example: "Resource not found"
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_FAILED"
            message:
              type: string
              example: "The given data was invalid."
            errors:
              type: array
              items:
                type: string
              example: ["template is required", "data must be an object"]

    # --- Connection & Health ---
...
    BranchListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Branch"
    Branch:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: "SDP-SBY"
        name:
          type: string
          example: "Sidoarjo Plant Surabaya"
        company_id:
          type: integer
...
    TemplateListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Template"
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "invoice_sewa"
        label:
          type: string
          example: "Invoice Sewa"
        description:
          type: string
        version:
          type: integer
        page_size:
          type: string
    TemplateDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/TemplateDetail"
    TemplateDetail:
      allOf:
        - $ref: "#/components/schemas/Template"
        - type: object
          properties:
            elements:
              type: array
              items:
                type: object
            dimensions:
              type: object
              properties:
                width_mm:
                  type: number
                height_mm:
                  type: number
...
    DocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Document"
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        label:
          type: string
        created_at:
          type: string
          format: date-time
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        meta:
          type: object
          properties:
            current_page:
              type: integer
            last_page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer

    # --- Connector Registry ---
...